+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.1.1
 
 *   Unified usage of template arugments in unit_t
+*   electric conductance can be built from 1/Ohm (including all ratios)
+*   electric resistance can be built from 1/S (including all ratios)
 
 ## 1.1.0